refactor: update tsconfig path aliases and refactor import paths
