+++ server/src/modules/analise-migracao/rules/exportacao-acima-limite.rule.ts
@@ -1,4 +1,4 @@
-import { Diagnostico } from '../analise-migracao.service';
+import { Diagnostico } from '@/interfaces/diagnostico';
 
 export function validaExportacaoAcimaLimite(
   user: any,
