+++ server/src/modules/analise-migracao/rules/quantidade-funcionario.rule.ts
@@ -1,4 +1,4 @@
-import { Diagnostico } from '../analise-migracao.service';
+import { Diagnostico } from '@/interfaces/diagnostico';
 
 export function validaQuantidadeFuncionarios(
   user: any,
