+++ server/src/modules/diagnostico-inicial/diagnostico-inicial.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
-import { DiagnosticoInicialService } from './diagnostico-inicial.service';
-import { DiagnosticoInicialController } from './diagnostico-inicial.controller';
+import { DiagnosticoInicialService } from '@/modules/diagnostico-inicial/diagnostico-inicial.service';
+import { DiagnosticoInicialController } from '@/modules/diagnostico-inicial/diagnostico-inicial.controller';
 import { HttpModule } from '@nestjs/axios';
-import { ReceitawsApiService } from 'src/integrations/receitaws-api/receitaws-api.service';
-import { AnaliseMigracaoModule } from '../analise-migracao/analise-migracao.module';
-import { AuthModule } from '../auth/auth.module';
+import { ReceitawsApiService } from '@/integrations/receitaws-api/receitaws-api.service';
+import { AnaliseMigracaoModule } from '@/modules/analise-migracao/analise-migracao.module';
+import { AuthModule } from '@/modules/auth/auth.module';
 
 @Module({
   imports: [HttpModule, AnaliseMigracaoModule, AuthModule],
