+++ server/src/modules/message/message.repository.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { PrismaService } from 'prisma/prisma.service';
+import { PrismaService } from '@prisma/prisma.service';
 import { CreateMessageDto } from './dto/create-message.dto';
 import { Mensagem } from '@prisma/client';
